Update workflow cd (#13)
* Update main.yml for CD

This commit updates main.yml to not only build and test the application
on push to main, but also build and publish a docker image to github
packages.

* Split CI/CD workflow

This commits separates the CI and CD part of the previous workflow in
two parts. main.yml reverts back to the build before adding the CD part.
docker_image.yml is a new workflow-file that builds and uploads a docker
image to github packages.