Merge development branch into main: Add dev workflow, modify handlers for dev testing (#19)
* Add workflow test_on_push_dev.yml

This commits adds the workflow file test_on_push_dev.yml.
The workflow will trigger on a push to a branch that starts with "dev".
The steps included is checking out the code, building and testing the
application.

* Modify handlers to test dev workflow

This commit alter the files handlers.go and handlers_test.go in purpose
to test the CI/CD workflow. This commit should the workflow
test_on_push_dev.yml when pushed to the development branch.