Update handlers.go and handlers_test.go
This commit updates the string in server/handlers.go and the
corresponding string in server/handlers_test.go
The string now contains information validating the latest release.

Furthermore the commit, when pushed to main, will trigger the workflow
create_publish_docker.yml after the tag 1.1.0 is pushed. This will
create and publish a docker image to github packages since a new tag is
being pushed.