Update handlers.go and handlers_test.go
This commit updates handlers.go and handlers_test.go to contain the
string "Hello, Nackademin! This is a new release!"

It also contains a fix for the create_publish_docker.yml workflow file.
New steps are added to set up Docker Buildx which is needed to build the
docker image.